Fix middleware namespace conflict
- fix test
