+++ lib/service_fusion_api/faraday/auth.rb
@@ -1,0 +1,15 @@
+module ServiceFusionApi
+  module FaradayMiddleware
+    class Auth < Faraday::Middleware
+      def initialize(app, access_token)
+        @app = app
+        @access_token = access_token
+      end
+
+      def call(env)
+        env[:request_headers].merge!("Authorization" => "Bearer #{@access_token}")
+        @app.call env
+      end
+    end
+  end
+end
