+++ lib/service_fusion_api/http_utils/response.rb
@@ -3,6 +3,7 @@ module ServiceFusionApi
     module Response
       def self.create(response_hash)
         data = response_hash['data'].dup rescue response_hash
+        data ||= response_hash
         data.extend(self)
         raise ServiceFusionApi::AuthorizationError.new(data.error) if data.respond_to?(:error)
         data
