+++ lib/service_fusion_api/version.rb
@@ -2,7 +2,7 @@ module ServiceFusionApi
   class Version
     MAJOR = 0 unless defined? ServiceFusionApi::Version::MAJOR
     MINOR = 1 unless defined? ServiceFusionApi::Version::MINOR
-    PATCH = 2 unless defined? ServiceFusionApi::Version::PATCH
+    PATCH = 3 unless defined? ServiceFusionApi::Version::PATCH
 
     class << self
       # @return [String]
